Recommit

+++ db/migrate/20120708230600_add_external_id_to_posts.rb
@@ -1,7 +1,7 @@
 class AddExternalIdToPosts < ActiveRecord::Migration
   def self.up
     add_column :posts, :external_id, :bigint
-    add_column :posts, :external_type, :tinyint
+    add_column :posts, :external_type, :smallint
     
     add_index :posts, [:user_id, :external_id, :external_type], :unique => true
   end
